requirements.txt
opencv-python==4.8.1.78
pyserial==3.5
pyyaml==6.0.1
python-dotenv==1.0.0
platformio==6.1.11